add weather to all xml show weather
