+++ app/src/main/java/com/example/planforplant/api/ApiClient.java
@@ -13,9 +13,9 @@ public class ApiClient {
     private static Retrofit plantNetRetrofit = null;
     private static Retrofit meteosourceRetrofit = null;
 
-    private static final String LOCAL_BASE_URL = "http://192.168.100.90:8080/";
+   // private static final String LOCAL_BASE_URL = "http://192.168.100.90:8080/";
     // nếu dùng điện thoại ảo
-//     private static final String LOCAL_BASE_URL = "http://10.0.2.2:8080/";
+    private static final String LOCAL_BASE_URL = "http://10.0.2.2:8080/";
     private static final String PLANTNET_BASE_URL = "https://my-api.plantnet.org/v2/";
     private static final String METEOSOURCE_BASE_URL = "https://www.meteosource.com/";
