+++ app/src/main/java/com/example/planforplant/ui/DetailActivity.java
@@ -2,14 +2,12 @@ package com.example.planforplant.ui;
 
 import android.content.Intent;
 import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.graphics.Matrix;
-import android.net.Uri;
 import android.os.Bundle;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.exifinterface.media.ExifInterface;
 
@@ -20,10 +18,9 @@ import com.example.planforplant.api.ApiService;
 import com.example.planforplant.model.Plant;
 import com.example.planforplant.model.PlantResponse;
 import com.example.planforplant.model.Result;
+import com.example.planforplant.weather.WeatherManager;
 import com.google.gson.Gson;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -34,35 +31,47 @@ import retrofit2.Response;
 public class DetailActivity extends AppCompatActivity {
     private ImageView plantImage;
     private TextView tvPlantName, tvOverview, tvFamily, tvGenus, tvSpecies;
-    private TextView tvKingdom, tvPhylum, tvClass, tvOrder;
+    private TextView tvPhylum, tvClass, tvOrder;
     private TextView tvWater, tvLight, tvTemperature, tvCareGuide, tvDiseases;
+
+    // Weather views
+    private TextView tvLocation, tvWeather;
+    private ImageView ivWeatherIcon;
+    private WeatherManager weatherManager;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.detail);
 
-        // Bind views
+        // Bind plant views
         plantImage = findViewById(R.id.plantImage);
         tvPlantName = findViewById(R.id.tvPlantName);
         tvOverview = findViewById(R.id.tvOverview);
         tvFamily = findViewById(R.id.tvFamily);
         tvGenus = findViewById(R.id.tvGenus);
         tvSpecies = findViewById(R.id.tvSpecies);
-
         tvPhylum = findViewById(R.id.tvPhylum);
         tvClass = findViewById(R.id.tvClass);
         tvOrder = findViewById(R.id.tvOrder);
-
         tvWater = findViewById(R.id.tvWater);
         tvLight = findViewById(R.id.tvLight);
         tvTemperature = findViewById(R.id.tvTemperature);
         tvCareGuide = findViewById(R.id.tvCareGuide);
         tvDiseases = findViewById(R.id.tvDiseases);
 
+        // Bind weather views
+        tvLocation = findViewById(R.id.tvLocation);
+        tvWeather = findViewById(R.id.tvWeather);
+        ivWeatherIcon = findViewById(R.id.ivWeatherIcon); // optional, if your layout has an icon
+
+        // Initialize WeatherManager
+        weatherManager = new WeatherManager(this, tvLocation, tvWeather, ivWeatherIcon);
+        weatherManager.start();
+
+        // Handle intent data
         Intent intent = getIntent();
 
-        // Case 1: CaptureActivity -> PlantResponse
         String json = intent.getStringExtra("plantResponseJson");
         if (json != null) {
             PlantResponse response = new Gson().fromJson(json, PlantResponse.class);
@@ -70,69 +79,63 @@ public class DetailActivity extends AppCompatActivity {
 
             if (response != null && response.results != null && !response.results.isEmpty()) {
                 Result first = response.results.get(0);
-
                 List<String> keywords = new ArrayList<>();
 
-                // Best match
-                if (response.bestMatch != null) {
-                    keywords.add(response.bestMatch);
-                }
+                if (response.bestMatch != null) keywords.add(response.bestMatch);
 
-                // Tên khoa học
                 if (first.species != null) {
                     if (first.species.scientificName != null)
                         keywords.add(first.species.scientificName);
-
                     if (first.species.scientificNameWithoutAuthor != null)
                         keywords.add(first.species.scientificNameWithoutAuthor);
-
-                    // Common names
                     if (first.species.commonNames != null)
                         keywords.addAll(first.species.commonNames);
                 }
 
-                // Genus + Family
                 if (first.species.genus != null && first.species.genus.scientificName != null)
                     keywords.add(first.species.genus.scientificName);
-
                 if (first.species.family != null && first.species.family.scientificName != null)
                     keywords.add(first.species.family.scientificName);
 
-                // Gọi backend thử lần lượt với từng keyword
                 searchWithFallback(keywords);
             }
             return;
         }
 
-        // Case 2: SearchResultsActivity -> Plant
         String plantJson = intent.getStringExtra("plantEntityJson");
         if (plantJson != null) {
             Plant plant = new Gson().fromJson(plantJson, Plant.class);
             bindPlantEntity(plant);
         }
     }
-        private void searchWithFallback(List<String> keywords) {
-            if (keywords == null || keywords.isEmpty()) return;
 
-            ApiService apiService = ApiClient.getLocalClient(this).create(ApiService.class);
+    @Override
+    public void onRequestPermissionsResult(int requestCode,
+                                           @NonNull String[] permissions,
+                                           @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        // Delegate permission results to WeatherManager
+        weatherManager.onRequestPermissionsResult(requestCode, permissions, grantResults);
+    }
 
-            // Duyệt lần lượt cho tới khi tìm thấy
-            for (String keyword : keywords) {
-                apiService.searchPlants(keyword).enqueue(new Callback<List<Plant>>() {
-                    @Override
-                    public void onResponse(Call<List<Plant>> call, Response<List<Plant>> response) {
-                        if (response.isSuccessful() && response.body() != null && !response.body().isEmpty()) {
-                            bindPlantEntity(response.body().get(0));
-                        }
+    // --- Plant data methods ---
+    private void searchWithFallback(List<String> keywords) {
+        if (keywords == null || keywords.isEmpty()) return;
+        ApiService apiService = ApiClient.getLocalClient(this).create(ApiService.class);
+        for (String keyword : keywords) {
+            apiService.searchPlants(keyword).enqueue(new Callback<List<Plant>>() {
+                @Override
+                public void onResponse(Call<List<Plant>> call, Response<List<Plant>> response) {
+                    if (response.isSuccessful() && response.body() != null && !response.body().isEmpty()) {
+                        bindPlantEntity(response.body().get(0));
                     }
+                }
 
-                    @Override
-                    public void onFailure(Call<List<Plant>> call, Throwable t) {
-                        // có thể log lại nhưng không stop
-                    }
-                });
-            }
+                @Override
+                public void onFailure(Call<List<Plant>> call, Throwable t) { }
+            });
         }
+    }
 
     private void bindPlantData(PlantResponse response) {
         if (response == null) return;
@@ -147,24 +150,19 @@ public class DetailActivity extends AppCompatActivity {
                 if (first.species.genus != null)
                     tvGenus.setText(first.species.genus.scientificName);
 
-                if (first.species.commonNames != null && !first.species.commonNames.isEmpty()) {
+                if (first.species.commonNames != null && !first.species.commonNames.isEmpty())
                     tvOverview.setText("Common names: " + String.join(", ", first.species.commonNames));
-                } else {
-                    tvOverview.setText("No common names available");
-                }
+                else tvOverview.setText("No common names available");
             }
         }
     }
 
     private void bindPlantEntity(Plant plant) {
         if (plant == null) return;
-
-        // Basic info
         tvPlantName.setText(plant.getCommonName() != null ? plant.getCommonName() : "Unknown");
         tvSpecies.setText(plant.getScientificName() != null ? plant.getScientificName() : "");
         tvOverview.setText(plant.getDescription() != null ? plant.getDescription() : "No description available");
 
-        // Taxonomy
         tvPhylum.setText(plant.getPhylum() != null ? plant.getPhylum() : "");
         tvClass.setText(plant.getPlantClass() != null ? plant.getPlantClass() : "");
         tvOrder.setText(plant.getPlantOrder() != null ? plant.getPlantOrder() : "");
@@ -172,16 +170,12 @@ public class DetailActivity extends AppCompatActivity {
         tvGenus.setText(plant.getGenus() != null ? plant.getGenus() : "");
         tvSpecies.setText(plant.getSpecies() != null ? plant.getSpecies() : "");
 
-        // Growth & care
         tvWater.setText(plant.getWaterSchedule() != null ? plant.getWaterSchedule() : "");
         tvLight.setText(plant.getLight() != null ? plant.getLight() : "");
         tvTemperature.setText(plant.getTemperature() != null ? plant.getTemperature() : "");
         tvCareGuide.setText(plant.getCareGuide() != null ? plant.getCareGuide() : "");
-
-        // Diseases (placeholder for now — backend chưa có trường này)
         tvDiseases.setText("🍂 Bệnh rụng lá sớm\n🕷️ Sâu đục thân\n🦠 Nấm mốc trắng");
 
-        // Load image if available
         if (plant.getImageUrl() != null && !plant.getImageUrl().isEmpty()) {
             Glide.with(this)
                     .load(plant.getImageUrl())
@@ -190,33 +184,6 @@ public class DetailActivity extends AppCompatActivity {
         }
     }
 
-    private void fetchPlantFromBackend(String keyword) {
-        ApiService apiService = ApiClient.getLocalClient(DetailActivity.this).create(ApiService.class);
-
-        apiService.searchPlants(keyword).enqueue(new Callback<List<Plant>>() {
-            @Override
-            public void onResponse(Call<List<Plant>> call, Response<List<Plant>> response) {
-                if (response.isSuccessful() && response.body() != null && !response.body().isEmpty()) {
-                    bindPlantEntity(response.body().get(0)); // ✅ show first plant
-                } else {
-                    // ❌ No result → NotFoundActivity
-                    Intent intent = new Intent(DetailActivity.this, NotFoundActivity.class);
-                    intent.putExtra("message", "Không tìm thấy cây trong cơ sở dữ liệu cho từ khóa: " + keyword);
-                    startActivity(intent);
-                    finish(); // close detail page
-                }
-            }
-
-            @Override
-            public void onFailure(Call<List<Plant>> call, Throwable t) {
-                // ❌ Error → NotFoundActivity with error message
-                Intent intent = new Intent(DetailActivity.this, NotFoundActivity.class);
-                intent.putExtra("message", "Lỗi khi gọi backend: " + t.getMessage());
-                startActivity(intent);
-                finish();
-            }
-        });
-    }
     private Bitmap rotateBitmap(Bitmap bitmap, int orientation) {
         if (bitmap == null) return null;
         Matrix matrix = new Matrix();
