+++ app/src/main/java/com/example/planforplant/weather/WeatherUtils.java
@@ -1,0 +1,50 @@
+package com.example.planforplant.weather;
+
+import com.example.planforplant.R;
+
+public class WeatherUtils {
+
+    /**
+
+     */
+    public static int getIconResource(int code) {
+        switch (code) {
+            case 2: return R.drawable.ic_2;
+            case 3: return R.drawable.ic_3;
+            case 4: return R.drawable.ic_4;
+            case 5: return R.drawable.ic_5;
+            case 6: return R.drawable.ic_6;
+            case 7: return R.drawable.ic_7;
+            case 8: return R.drawable.ic_8;
+            case 9: return R.drawable.ic_9;
+            case 10: return R.drawable.ic_10;
+            case 11: return R.drawable.ic_11;
+            case 12: return R.drawable.ic_12;
+            case 13: return R.drawable.ic_13;
+            case 14: return R.drawable.ic_14;
+            case 15: return R.drawable.ic_15;
+            case 16: return R.drawable.ic_16;
+            case 17: return R.drawable.ic_17;
+            case 18: return R.drawable.ic_18;
+            case 19: return R.drawable.ic_19;
+            case 20: return R.drawable.ic_20;
+            case 21: return R.drawable.ic_21;
+            case 22: return R.drawable.ic_22;
+            case 23: return R.drawable.ic_23;
+            case 24: return R.drawable.ic_24;
+            case 25: return R.drawable.ic_25;
+            case 26: return R.drawable.ic_26;
+            case 27: return R.drawable.ic_27;
+            case 28: return R.drawable.ic_28;
+            case 29: return R.drawable.ic_29;
+            case 30: return R.drawable.ic_30;
+            case 31: return R.drawable.ic_31;
+            case 32: return R.drawable.ic_32;
+            case 33: return R.drawable.ic_33;
+            case 34: return R.drawable.ic_34;
+            case 35: return R.drawable.ic_35;
+            case 36: return R.drawable.ic_36;
+            default: return R.drawable.ic_1;
+        }
+    }
+}
